rejigged to use QUnit testing framework and CommonJS require()
The require() function is meant to be a cross-platform way of loading another javascript file.
If we replace all the rhino-specific stuff with CommonJS functions, the tests should be able to run on any javascript environment.

I couldn't work out how to get the CommonJS filesystem module loaded, so we're stuck with a call to readFile, which node.js doesn't have.

The tests now run through QUnit, which is a mature testing framework developed by the people behind jQuery.

I haven't changed the format of the yaml tests, but I've moved them to standalone-jme/tests/, and added a bit of support for having multiple test spec files.